+++ mywords-flutter/lib/pages/article_archived_list.dart
@@ -1,9 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:mywords/common/queue.dart';
-import 'package:mywords/libso/handler.dart';
-
 import 'package:mywords/widgets/article_list.dart';
-
 import 'package:mywords/util/get_scaffold.dart';
 
 class ArticleArchivedPage extends StatefulWidget {
@@ -24,7 +21,6 @@ class _State extends State<ArticleArchivedPage> {
           actions: [
             IconButton(
                 onPressed: () {
-                  // 归档置顶
                   produceEvent(EventType.articleListScrollToTop, 2);
                 },
                 icon: const Icon(Icons.vertical_align_top_outlined))