+++ mywords-flutter/lib/pages/share_data.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:mywords/libso/handler.dart';
@@ -6,7 +5,6 @@ import 'package:path_provider/path_provider.dart';
 import 'package:path/path.dart' as path;
 import 'package:mywords/widgets/private_ip.dart';
 
-import 'package:mywords/libso/resp_data.dart';
 import 'package:mywords/util/get_scaffold.dart';
 import 'package:mywords/util/util.dart';
 import 'package:url_launcher/url_launcher_string.dart';
@@ -119,14 +117,6 @@ class _SyncDataState extends State<SyncData> {
     List<Widget> children = [const PrivateIP()];
 
     children.add(ListTile(
-      leading: Tooltip(
-        // message:            "开启后将允许其他设备访问进行同步本机数据，也可以在浏览器中进行下载 http://ip:port/code,\n例如 http://$localExampleIP:${shareInfo.port}/${shareInfo.port}",
-        message:
-            "After opening, other devices will be allowed to access and synchronize local data, and you can also download it in the browser http://ip:port/code,\nfor example http://$localExampleIP:${shareInfo.port}/${shareInfo.port}",
-        triggerMode: TooltipTriggerMode.tap,
-        showDuration: const Duration(seconds: 30),
-        child: const Icon(Icons.info_outline),
-      ),
       title: Row(
         children: [
           Expanded(
@@ -167,11 +157,12 @@ class _SyncDataState extends State<SyncData> {
       trailing: switchBuild(),
     ));
     if (shareInfo.open) {
+      final hostname = handler.getHostName();
       final shareFileInfosURL =
-          "http://127.0.0.1:${shareInfo.port}/share/shareFileInfos?code=${shareInfo.code}";
+          "http://$hostname:${shareInfo.port}/share/shareFileInfos?code=${shareInfo.code}";
 
       children.add(ListTile(
-        title: const Text("文件列表"),
+        title: const Text("File List"),
         leading: const Icon(Icons.http),
         subtitle: Text(shareFileInfosURL),
         trailing: IconButton(
@@ -182,10 +173,10 @@ class _SyncDataState extends State<SyncData> {
       ));
 
       final shareKnownWordsURL =
-          "http://127.0.0.1:${shareInfo.port}/share/shareKnownWords?code=${shareInfo.code}";
+          "http://$hostname:${shareInfo.port}/share/shareKnownWords?code=${shareInfo.code}";
       children.add(ListTile(
         leading: const Icon(Icons.http),
-        title: const Text("Known Words"),
+        title: const Text("My Words Library"),
         subtitle: Text(shareKnownWordsURL),
         trailing: IconButton(
             onPressed: () {