+++ mywords-flutter/lib/widgets/article_list.dart
@@ -66,19 +66,18 @@ class _State extends State<ArticleListView> {
     final itemNew = item.copyWith(archived: false);
     final t = Timer(const Duration(milliseconds: 3500), () async {
       final respData = await handler.updateFileInfo(itemNew);
+      produceEvent(EventType.updateArticleList);
       if (respData.code != 0) {
         myToast(context, respData.message);
         return;
       }
-      produceEvent(EventType.updateArticleList);
     });
     // Then show a snackbar.
     ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-      // content: Text('文章已取消归档: ${item.title}',
       content: Text('Article has been unarchived: ${item.title}',
           maxLines: 1, overflow: TextOverflow.ellipsis),
       action: SnackBarAction(
-          label: "撤销",
+          label: "Revoke",
           onPressed: () {
             t.cancel();
             initFileInfos();
@@ -91,6 +90,8 @@ class _State extends State<ArticleListView> {
     final itemNew = item.copyWith(archived: true);
     final t = Timer(const Duration(milliseconds: 3500), () async {
       final RespData respData = await handler.updateFileInfo(itemNew);
+      produceEvent(EventType.updateArticleList);
+
       if (respData.code != 0) {
         if (!mounted) {
           return;
@@ -98,14 +99,13 @@ class _State extends State<ArticleListView> {
         myToast(context, respData.message);
         return;
       }
-      produceEvent(EventType.updateArticleList);
     });
-    // Then show a snackbar.
+    // Then show a SnackBar.
     ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-      content: Text('文章已归档: ${item.title}',
+      content: Text('Article has been archived: ${item.title}',
           maxLines: 1, overflow: TextOverflow.ellipsis),
       action: SnackBarAction(
-          label: "撤销",
+          label: "Revoke",
           onPressed: () {
             t.cancel();
             initFileInfos();
@@ -118,19 +118,19 @@ class _State extends State<ArticleListView> {
     final id = item.id;
     final t = Timer(const Duration(milliseconds: 3500), () async {
       final RespData respData = await handler.deleteGobFile(id);
+      produceEvent(EventType.updateArticleList);
       if (respData.code != 0) {
         if (!mounted) return;
         myToast(context, respData.message);
         return;
       }
-      produceEvent(EventType.updateArticleList);
     });
     // Then show a snackbar.
     ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-      content: Text('文章已删除: ${item.title}',
+      content: Text('Article has been deleted: ${item.title}',
           maxLines: 1, overflow: TextOverflow.ellipsis),
       action: SnackBarAction(
-          label: "撤销",
+          label: "Revoke",
           onPressed: () {
             t.cancel();
             initFileInfos();
@@ -174,7 +174,7 @@ class _State extends State<ArticleListView> {
             background: getBackgroundWidget(
               context,
               left: getBackgroundChild(widget.leftLabel, widget.leftIconData),
-              right: getBackgroundChild("删除", Icons.delete),
+              right: getBackgroundChild("Delete", Icons.delete),
             ),
             direction: DismissDirection.horizontal,
             onDismissed: (DismissDirection direction) {
@@ -228,9 +228,7 @@ class _State extends State<ArticleListView> {
         context: context,
         builder: (BuildContext context) {
           return AlertDialog(
-              // title: const Text("提示"),
               title: const Text("Tips"),
-              // content: const Text("向左滑动删除文章, 向右滑动归档文章"),
               content: const Text(
                   "Swipe left to delete the article, swipe right to archive the article"),
               actions: [