+++ CHANGE_LOG.md
@@ -6,6 +6,8 @@
 - Support copy of Email and copy of build information in about dialog.
 - Upgrade dependencies to the latest version in go.mod.
 - Remove the redundant code in the project. Optimize Makefile.
+- Sensitive functions like `DelDict` are not allowed to be called in the web version when operating not in the local environment.
+- Net proxy settings support configuration of username and password.
 # v3.0.0
 - The underlying code has been refactored to make it clearer. Using `sqlite` to store data, making the project more stable and easier to maintain.
 - Mobile/desktop client version directly integrates Web, seamless learning between multiple devices. It called `Web Online` in the client version.